# [ONNX] Limit number of elements to display for list/tuple/dict in diagnostics