--- a/torch/onnx/_internal/diagnostics/infra/formatter.py
+++ b/torch/onnx/_internal/diagnostics/infra/formatter.py
@@ -65,7 +65,7 @@ def _convert_key(
         else:
             new_v = v
         if new_v is None:
-            # Otherwise unnesseraily bloated sarif log with "null"s.
+            # Otherwise unnecessarily bloated sarif log with "null"s.
             continue
         if new_v == -1:
             # WAR: -1 as default value shouldn't be logged into sarif.