--- a/torch/onnx/_internal/fx/diagnostics.py
+++ b/torch/onnx/_internal/fx/diagnostics.py
@@ -21,6 +21,9 @@ from torch.onnx._internal.fx import registration, type_utils as fx_type_utils
 # Tensor(i64[s0, 64, (s1//2) - 2, (s1//2) - 2]) where s0 and s1 are symbolic
 # so we need to relax the length limit.
 _LENGTH_LIMIT: int = 120
+# NOTE: The following limits are for the number of items in a list, tuple or dict.
+# The limits are set to be small to avoid heavy workload.
+_CONTAINER_ITEM_LIMIT: int = 10
 
 # NOTE(bowbao): This is a shim over `torch.onnx._internal.diagnostics`, which is
 # used in `torch.onnx`, and loaded with `torch`. Hence anything related to `onnxscript`
@@ -132,7 +135,11 @@ def _list(obj: list) -> str:
     list_string = f"List[length={len(obj)}](\n"
     if not obj:
         return list_string + "None)"
-    for item in obj:
+    for i, item in enumerate(obj):
+        if i >= _CONTAINER_ITEM_LIMIT:
+            # NOTE: Print only first _CONTAINER_ITEM_LIMIT items.
+            list_string += "...,\n"
+            break
         list_string += f"{format_argument(item)},\n"
     return list_string + ")"
 
@@ -142,7 +149,11 @@ def _tuple(obj: tuple) -> str:
     tuple_string = f"Tuple[length={len(obj)}](\n"
     if not obj:
         return tuple_string + "None)"
-    for item in obj:
+    for i, item in enumerate(obj):
+        if i >= _CONTAINER_ITEM_LIMIT:
+            # NOTE: Print only first _CONTAINER_ITEM_LIMIT items.
+            tuple_string += "...,\n"
+            break
         tuple_string += f"{format_argument(item)},\n"
     return tuple_string + ")"
 
@@ -152,7 +163,11 @@ def _dict(obj: dict) -> str:
     dict_string = f"Dict[length={len(obj)}](\n"
     if not obj:
         return dict_string + "None)"
-    for key, value in obj.items():
+    for i, (key, value) in enumerate(obj.items()):
+        if i >= _CONTAINER_ITEM_LIMIT:
+            # NOTE: Print only first _CONTAINER_ITEM_LIMIT items.
+            dict_string += "...\n"
+            break
         dict_string += f"{key}: {format_argument(value)},\n"
     return dict_string + ")"